requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
evaluate>=0.4.0
seqeval>=1.2.2
accelerate>=0.27.0

# Document AI
Pillow>=10.0.0
pytesseract>=0.3.10       # OCR fallback


# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# UI
gradio>=4.20.0

# LLM QA
google-generativeai>=0.5.0

# Data / Eval
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
